Fix AST dependency computation.
Turns out the logic wasn't quite right yet. It worked well enough if
all units get wrapped into `ValueReference` because then we just need
a forward declaration to make that work; and those we would always
output first. But it wouldn't actually track dependencies correctly to
output them in the right order so that full struct are available where
needed. This commit reworks that logic. Turns out this also further
simplifies logic and state tracking, which is a good sign.

By itself, because everything was already working, this fix doesn't
change much other than some baselines now showing different order for
some declarations as well as more dependencies in debug outout that we
had missed earlier. But I have tested it with the `on-heap` branch
that moves structs to the stack, for which forward declarations are no
longer sufficient, and that branch now works correctly, whereas it
didn't before.