Simplify parsing of literals.
There are two changes in here:

1. For grammars that don't use look-ahead, we skip the runtime check for a
pending look-ahead symbol, because we know we will never have one. This removes
generated code of the form `if ( _lah ) { ... }`. This change needs a bit of
machinery unfortunately because we need to get the information about look-ahead
usage over into the codegen for literal parsing.

2. For bytes literals, we now outsource their parsing to a runtime function to
make the generated code simpler. As a side effect this also provides more
informative error messages when the literal isn't found.

Taking the two together means that the code for parsing a plain `b"Foo"`
literal may look like this now:

    spicy_rt::expectBytesLiteral(__data, __cur, b"Foo", "b.spicy:4:10-4:15", Null);
    __cur = __cur.advance(3);

    if ( __trim )
        (*__data).trim(begin(__cur));

    (*self).foo = b"Foo";