Add peephole optimizer for final AST tuning.
We use this to remove two statement constructs that the main optimizer
may leave behind:

    1. `default<void>()`
    2. `(*self).__error = __error; __error = (*self).__error;`

The second case is a quite specific situation that eventually, once we
have CFG/DFG tracking, the main optimizer should be able to cover more
generically. However, for now, it's just not nice to always have
these blocks in the generated C++ code, so adding this special case
seems useful.

Couples notes on (2):

    - Per #1592, case 2 may also have overhead. Closes #1592.
    - Technically, this optimization isn't always correct: subsequent
      code could assume that `(*self).__error` is set, whereas after
      removal it's not (or not to the expected value). However,
      `__error` is strictly-internal state, and we know that we don't
      use it any different, so this seems ok until we have more
      general optimizer logic.