Add support for field-local `%error` handlers.
We now support attaching an `%error` handler to an individual field:

    type Test = unit {
        a: b"A";
        b: b"B" %error { print "field B %error", self; }
        c: b"C";
    };

With input `AxC`, that handler will trigger, whereas with `ABx` it
won't. If the unit had a unit-wide `%error` handler as well, that one
would trigger in both cases (i.e., for `b`, in addition to its field
local handler).

The handler can also be provided separately from the field:

    on b %error { ... }

In that separate version, one can receive the error message as well by
declaring a corresponding string parameter:

    on b(msg: string) %error { ... }

This works externally, from outside the unit, as well:

    on Test::b(msg: string) %error { ... }

This is eebased on top of `topic/robin/optimize-type-parsing` so that
we get the peephole optimizer.

Addresses #1824.