Add a new unit item `Block` that stores a sequence of subitems.
We now have a new unit item `Block` that represents a sequence of
sub-items. This goes along with a new production `Block` that the item
gets turned into inside the grammar.

The implementation reuses logic from `unit::item::switch_::Case`. We
then refactor `Case` to use the new `Block` instead of continuing to
maintain its own list of items.

We also refactor some other logic
previously located with `Switch` to now apply to `Blocks` as well.

Optionally, a `Block` can have a conditional expressions, an
"else"-block as well attributes with it. We don't use them yet, but
will soon; tests then coming as well. However, we already also
refactor some attribute logic that previously went just with the
`Switch` unit item to now also apply to the `Block` item, so that the
two can support the same attributes.