Add internal error for unvalidatable vectors.
Vectors were changed to never be a vector-of-type, just a
vector-of-items that redirect to a type. This allows validating
attributes and whatnot properly, which only apply to fields, not types.
The consequence is that a vector-of-types should be impossible, or else
there is an error in the parser's workarounds to get to that state,
which is a bug.

Add an internal error so that case never surfaces without a big scary
error.