Provide proper error message when trying access an unknown unit field.
Returning `type::Unknown` instead of `type::Auto` will let the
resolver process stop, allowing an already existing error message to
kick in later. We were doing it this way already for structs, but not
for units.

This also includes a fix for a bug with finding bitfield ranges by ID,
which was triggered by the change.

Closes #1790.