Fix usage of `&convert` with unit's requiring parameters.
The generated code needs to create a temporary instance of the type,
but doesn't have any arguments to provide to it. But that's ok, and we
now let the validation pass and just instantiate a default-constructed
instance.

However, this change now requires an additional validator check on the
Spicy side to ensure fields giving arguments to types do so correctly
Before we happened to check that implicitly on the HILTI-side through
code that now would let it pass if no arguments were given.

Closes #1791.