Add infrastructure to create and cache global constants.
So far we had replicated low-level code a few times that cached
constants through a global declaration for reuse. This now factors out
that logic into a central method inside the Spicy code generator.

This isn't full interning, which I'm not sure we want/need, but covers
the performance use case that has now come up a few times.