Move type alias attribute validation to parser.
Closes #1898

This was originally in a validator, but there's some sense that this is
a parsing issue since attributes should not be allowed at all on most
constructs of that form.

Note that a lot of the cleanup here was meant to span removing the
`opt_attributes` field in `type_decl` altogether, but the fallout from
that is pretty large. It ended up being less cleanup and more
rearchitecting. This way still avoids the hacky test for type aliases in
the validator while not changing the AST structure.