Add new "condition-test" operator.
This adds a new expression operator to both HILTI and Spicy:

    `COND : ERROR`

where `COND` is a boolean expression, and `ERROR` is an expression of
type `error`. When evaluated, the new operator yields a value of type
`result<void>`, which will be true if the `COND` is true, and set to
`ERROR` if `COND` is false. In other words, this is a short-cut to
both test a condition and provide an error message in case it fails.