Teach `&requires` to accept condition-tests expressions.
This now allows creating custom error messages when a `&require`
condition fails. Example:

    type Foo = unit {
        x: uint8 &requires=($$ == 1 : error"Deep trouble!'");

        # or, shorter:
        y: uint8 &requires=($$ == 1 : "Deep trouble!'");
    };

Internally, `&requires` now always wants a condition-test expression.
That remains transparent to the user, however, because the resolver
knows how to coerce a pure boolean `&requires` expression into a
condition test, crafting an error message that happens to look like
what we used to produce in the past.

Closes #1856.