Add consistent validation for attributes.
Closes #1901

This updates attribute validation to use some fancy bitmask magic in
order to see if attributes are in the right places. The goal here is to
have one place (ie a big map) to answer the question "Can I use X
attribute on Y node?"

There are a lot more moving parts with attribute validation, but those
generally have to do with behavior. A lot of that requires extra context
in the validator, which is exactly what the validator is meant to do.
Much of that is pretty ad-hoc, so it could get cleaned up as well.