Allow defining parser alias names when running spicy-driver.
``--parser-alias <alias>=<name>`` now makes the parser of name
`<name>` accessible under an additional name `<alias>`. The option can
be used multiple times.

This is particularly useful when using input batches: by adding an
alias of the form `PORT/PROTOCOL` (e.g., `--parser-alias
80/tcp=MyHTTP::Message`), this allows to define a parser to use on for
connections on that port (because the batch processing interprets any
ports in the trace as strings and tries to lookup a parser of the
corresponding name). Adding `%orig` or `%resp` to the alias name
limits it to the corresponding direction (e.g., ``--parser-alias
80/tcp%orig=MyHTTP::Request``).

See the `spicy-driver` documentation for more details on how to use
this with batches.