Deprecate `&count` attribute.
The preferred way to indicate how many elements should be parsed for a
vector has been vector syntax over `&count` for a long time, e.g.,

```
type X = unit {
    : uint8[] &count=42; # AVOID.
    : uint8[42];         # PREFER.
};
```

With this patch `&count` is now deprecated and we emit a warning. We
will remove support for `&count` in a future release.

Closes #1938.