Make sure autogen-docs pre-commit hook can always run in CI.
We would previously exit early from this script if we could not generate
the architecture diagram for which we needed the Python `diagrams`
package (which implicitly depends on `dot` as well). This setup made it
impossible to automatically update generated documentation unless one
had both `diagrams` and `dot` installed. Additionally the generated
`architecture.svg` was not stable across different versions of `dot`.

This patch rewrites `./doc/scripts/autogen-docs` so it can update
documentation even if the dependencies for diagram generation are
present. Since the `diagrams` output depends on dot (and its version) I
currently see no stable, platform-indepdendent way to generate the
diagram in a pre-commit hook, so we do not even attempt to make
`diagrams` available to the pre-commit hook.

We also fix a few other issues in the generator script:

- make sure it works when `gsed` is not present
- make sorting order of `reserved-keywords.txt` determistic across GNU
  and BSD toolchains