Enable productions to provide the number of bytes they parse, if known.
For productions that parse a static amount of data, that size can now
be retrieved. This includes cases where the number is determined
through some expression evaluated at runtime, as long as parsing the
production is guaranteed to consume always exactly as many bytes as
the expression yields. The new functionality will be used, and tested,
in subsequent commits.

This also refactors the implementation of `skip` to make use of the
same, new machinery to determine the number of bytes a production
consumes.