Fix `ParserBuilder::advanceInput()`.
The implementation tried to automatically detect if it was passed a
view or something else. However, that check was unreliable because the
type of the argument could still be unresolved (i.e., `auto`). This
change removes that broken support for passing a view; seems that
wasn't actually used anywhere anyways (probably precisely because it
didn't work).