Fix codegen for string literals containing null bytes.
Our strings are UTF8 which can contain literal null bytes. We were
previously generating incorrect C++ for such literals. This was due to
us constructing C++ value via constructors taking `const char*` which is
expected to be null-terminated, i.e., everything after the termiantor
was ignored even though it was emitted.

With this patch we switch to C++ literals for generating literal and
runtime strings.

Closes #1949.