Proof-of-concept size-based synchronization at top-level.
This aims to address a situation like this:

    type Chunks = unit {
        chunks: (Chunk &synchronize)[];
    };

    type Chunk = unit {
        content_size: bytes &until=b"\n" &convert=$$.to_uint();
        content: bytes &size=self.content_size;
    };

If an error happens while parsing `content` (e.g., a gap in the
input), the top-level `chunks` should be able to just move on the
subsequent chunk because it's clear where that start (namely after
`contents_size` bytes).

This proof-of-concept implementation makes this example work, but
isn't very pretty nor tested in detail; and other tests fail
currently. Need to clean up and re-assess.