Used fixed initialization for order for global static.
The `HILTI_PRE_INIT` macros work on the global
`_registered_preinit_functions` container. Since uses of the macro and
the global container are likely in different translation units their
initialization order can be undetermined (even though we have some
manual checks whether the `unique_ptr<vector>>` is null).

This patch cleans up initialization of the global so it should always be
fully initialized when accessed.