Provide generated, C++-side structs and unions with explicit `__to_string` methods.
This replaces the `__visit` methods we had so far, which we only used
for implementing the free-standing ``to_string()`` overloads for structs and units.

This is primarily cleanup, simplifying our generated code by avoiding
the hard-to-read `__visit` indirection, and maybe also speeding up
compilation times a bit because we no longer need any templates.

We also rename the `__str__` method (used for implementing the `on
%print` hook) to `__hook_to_string` for consistency.

Note that external C++ code should continue to call the free-standing
`to_string()` function to render a struct/unit, not the new method,
because the method doesn't take the `on %print` hook into account.