Fix coercion of function call arguments.
Before, we could unnecessarily coerce integer types when a direct
match was available. The fix takes out the `TryCoercionWithinSameType`
style for non-coercing function call arguments, which solves it. I
don't recall why that was in there in the first place, but it doesn't
seem to break anything to skip it ...

Closes #1968.