Avoid creating external types unnecessarily during AST construction.
We could end up creating external types for freshly created AST nodes
just because we wanted them to have a certain constness/sideness. This
could then lead to the original types not being validated because they
weren't inserted into the AST anywhere.

This also extends the AST dump output for qualified types to indicate
whether it's an internal or external type.