Switch regular expressions to new justrx API.
justrx no longer parses group match IDs out of the expression string,
but instead now offers an API to set them directly. This switches over
to using that new API.

As a result, setting a match ID inline inside a regexp doesn't work
anymore. An example of old syntax was: `/foobar${#100}/`. Instead, we
now allow IDs to be provided after the final slash. The new syntax is
`/foobar/$(100)`. While this is technically a breaking change, I doubt
anybody has been using this, and I don't think it's worth putting any
backwards-compatibility in.

Admittedly, that new ID syntax is bit arbitrary, and mostly chosen so
that it can be parsed easily. But it's not something users will
commonly use, so that seems fine. The Bison parser for that new syntax
already prepares for more pattern flags as well that we will want to
support.