Reject function overloads which only differ in default'd parameters.
Resolution of calls to functions who's signatures differ only in
defaulted parameters behaved non-intuitively in that when given with
less than the maximum possible number of parameters we resolved to the
overload with most parameters.

With this patch we reject such overloads where the only difference is in
defaulted parameters.

Closes #1981.