Change FFI calling convention for `any`.
When translating HILTI functions to C++, we now convert any HILTI-side
arguments of type `any` into *two* C++-side arguments: the value (as
before) plus a `TypeInfo*` pointer describing the value's type.

Before, we did this manually in a few cases where we needed the type
information by adding an additional `type` parameter to the HILTI
function. Now this happens implicitly and is therefore automatically
available to all functions using `any` parameters. The advantage is
that we don't need to modify an call sites to provide type information
explicitly, in particular when they are inside generated code. The
downside is that we end up receiving that new parameter in a few cases
where we don't actually need it (though this could change in the
future if we ever started to move away from templates inside the C++
runtime).