Remove one `__result` return indirection.
There are a few places a `__result` temporary is used to store the
return value. This should be pretty easy to optimize, but for some
reason even simple cases are slow to run. This takes one case and
simplifies the code.

Note that there is another case that a `__result` temp is created and
that also gives speedup when removed. That, however, has a possible
`finally` call after, and has a feature guard, so it's hard to optimize
that in the parser builder.

That case may be best with control flow optimization down the line.
Since this case is easy, it seems easier to debug if we can just
generate less code when the other code isn't necessary.