Add mechanism to leave temporary codegen variables uninitialized.
Local variables now support a "special" init expression for
performance optimization: If the init expression is an instance of
`expression::Void`, the generated C++ code will not default-initialize
the variable. That means it's not safe to read from it before it has
been written to at least once. This avoids the overhead of the
creating the default value when it's not needed, but it disables
HILTI's safety property of forcing all runtime values to have
well-defined content at all times.

This is a work-around until we have flow-based optimization which
should be able to figure out automatically when it's safe to skip
computing the default.