Fix using `&try` with `&max-size`, and potentially other cases.
In the generated code, a temporary was declared inside the `try { ...
}`  block that we create for `&try`, but then used afterwards outside
of the block. A similar issue probably could have happened in
other situations as well.

The fix is to move more of the logic that was so far following the
`try`-block into the block. This comes with potential slight semantic
changes where in case of a parse error, before we would have executed
some post-field logic that now skip. The main example is hooks:
before, if a field with a `&try` had a hook, this would still run in
case of a parse error; now it's not running anymore—--which indeed
seems the better semantics.

Closes #2007.