Fix `&size` expressions evaluating multiple times.
Fixes #2016

If the attribute is evaluated multiple times, some cases with side
effects may unknowingly get evaluated multiple times. This effect would
be different depending what happens - if there's a parse error, it could
even trigger an integer overflow when calculating offsets.

This solution uses a temporary and passes it around the parser builder
when it's needed. The downside is this is only applicable to `&size`,
but, luckily, that seems to be all that it currently applies to.

There are some alternative solutions, but they don't seem right. One
would be adding the temporary to the attribute, but that relies on only
retrieving that temporary if you're in the correct scope. Making the
attribute itself scope-aware seems wrong, as does removing the stored
expression when it would be out of scope.

It could also technically apply to the state, but that ended up being
more cumbersome than seemed necessary.