plit attributes between HILTI and Spicy.
A previous change had moved knowledge of Spicy attributes into the
HILTI layer. This splits that out again to separate concerns. The main
challenge is that we were using C++ enums, which could not be extended
by Spicy to add its own attributes. We now replace those enums values
with a custom helper class so that we can define constants at
multiple places.

Other than that new helper class, most of the changes are pretty
mechanical, with the main objective being that Spicy only refers to
its own attributes for anything that can appear in Spicy source code,
and uses HILTI attributes only when interacting with the HILTI layer.

Closes #1955.