Only disable clang-tidy if CMake version supports it.
We use `SKIP_LINTING` to disable running tools like `clang-tidy` via
CMake on generated. This is only supported with >=cmake-3.27 which is
satisfied in our CI infrastructure.

It is not really clear to me how to properly do this with earlier CMake
versions, so we simply only set this up if supported. This means that
users trying to run linters via CMake with an older version might see
diagnostics in generated code. This seems acceptable to me.