Support synchronization points at fixed offsets, without needing pattern search.
If a synchronization point (i.e., a field with `&synchronize`) can be
determined to reside at a fixed offset from either the start of a unit
or a previous synchronization point, we can now leverage that after an
error to jump there directly to resume parsing.

Note that this may change semantics for existing units: If there's a
`&synchronized` that used to operate by token searching, but now
qualifies for offset-based synchronization, the latter is preferred.
We see this actually in a number of existing synchronization tests,
which we update to force token-based matching so that we continue to
test the same functionality.