Support size-based synchronization inside containers.
This addresses a situation like this:

    type Chunks = unit {
        chunks: (Chunk &synchronize)[];
    };

    type Chunk = unit {
        content_size: bytes &until=b"\n" &convert=$$.to_uint();
        content: bytes &size=self.content_size;
    };

If an error happens while parsing `content` (e.g., a gap in the
input), the top-level `chunks` should be able to just move on the
subsequent chunk because it's clear where that starts (namely after
`contents_size` bytes).

This needs a bit more cleanup, more testing, and potentially support
for other cases.

Closes #1135.