Drop unnecessary Bison include from CMake function.
We do not seem to need any includes from Bison, but where previously
adding some include path in a pretty ad hoc way. This patch performs
minimal cleanup of this. There is probably the opportunity to clean up
how we add dependencies with Flex and Bison (e.g., using by refactoring
how we propagate this and using functionality like
`add_flex_bison_dependency` at the lowest level), but I punt on that
here.