Add workaround for rendering init of locals with void expressions.
We use void expressions as a hint for the initialization mechanism to
use in C++ codegen. These expressions have no actual equivalent in the
syntax.

To still somehow capture them in HILTI output render them by declaring a
variable with no constructor arguments; this is distinct from HILTI
default initialization which uses assignment syntax. This is not 100%
equivalent, but allows rendering valid code.