Allow disabling building of tests.
This patch adds a new CMake option `SPICY_ENABLE_TESTS` which if toggled
on forces building of test and benchmark binaries; it is `ON` by
default. This flag can be used by projects building Spicy to disable
building of tests if they are not interested in them. We also provide a
configure flag `--disable-tests` which has has the effect of turning it
off.