Prevent throwing naked exception when yielding from aborted fiber.
It is not clear to me why we did not throw a `RuntimeError` in the first
place, but in any case we did not provide a proper stringification which
would lead to a very cryptic error for this already obscure error
scenario.

With this patch we instead throw a Runtime error with some context.

(cherry picked from commit 647a3d94bf506ba5ca9f982722013c895bdfd617)