Fix fiber abortion.
When aborting a fiber, we need to activate it once more, to then leave
it for good by raising an `AbortException. Problem was that that
exception ended up being caught by user code because it was derived
from `std::exception`. This change removes the base class so that the
exception is guaranteed to go back to the managing fiber code, where
we just ignore it.