Speed up checking of iterator compatibility.
We were previously using a control block which held a weak_ptr to the
protected data. This was pretty inefficient for a number of reasons:

- access to the controlled data always required a `weak_ptr::lock` which
  created a temporary shared_ptr copy and immediately destroyed it after
  access
- to check whether the control block was expired we used `lock` instead
  of `expired` which introduced the same overhead
- to check compatibility of iterators we compared shared_ptrs to the
  control data which again required full locks instead of using
  `owner_before`

This patch introduces a new control block data structure and uses it
across all classes which previously held ad hoc implementations
(`Bytes`, `Map`, `Set`, `Vector`). The main improvement is that we now
separate tracking liveliness and the data, and use a better
implementation for control block equality checks. With the new
implementation I see throughput improvements across the board for
anything needing to iterate, e.g., I see bytes iteration being up to 30x
faster in trivial setups; the code from the original issue is now 10x
faster.

Closes #1663.