Track HILTI operator uses.
Closes #2029

This does not provide much rewriting infrastructure yet, but the rest
can be handled with helpers as needed. One use would be provide a
function which writes something before or after a given call (which may
be in a block). This would be necessary if we remove parameters that may
have side effects in the future. It's also necessary later when changing
return types and values of functions. But, for now, it is not.