Remove unused function parameters.
Closes #2030

This uses the new operator uses in order to remove function parameters.
Technically this can be done within the visitor, but it sets up to not
require visiting all expressions when pruning decls and uses. It's also
conceptually simpler (to me).

This will be particularly useful after a couple more optimizations,
where parse functions won't need nearly as many arguments if they're
only used in returns (and simply passed through without ever using
them). This is a relatively common pattern, enough so that this should
decrease the overhead of large numbers of small units substantially when
combined with future optimizations.