When C++ compilation fails, ask user for help.
This records the C++ compiler in a dedicated output file
"hilti-jit-error.log", and asks the user to file a ticket. If the
environment variable `HILTI_JIT_SHOW_CXX_OUTPUT` is set, the output
will also go to stderr. We set this variable inside CI.

Before, the process management code was trying to capture the output
in memory, but that didn't actually work, it always ended up on stderr
anyways. Now we tell reproc to just write all compiler output into a
temporary file, where we can then copy it from if needed.

Closes #2066.