Render all bitfield instances with included field names.
This requires passing the bitfield type information into bitfield
instances, which then allows us to unify the printing.

Most of the change is pretty mechanical to get us access to the
typeninfo. Plus, this also contains a fix where the `bits` typeninfo
previously would point directly to the stored integer type, even
though we wrap the integers into `optional` values and hence the
typeninfo should correctly reflect that.