Allow registering multiple modules in the same scope.
In 2ac6fa117615662472a0b578a991e76fc582a518 we switched to dynamically
injecting scopes for libraries. This assumed that
`Library::currentScope` was called exactly once per collection of
inputs, but this is not the case if there are multiple modules in such a
collection which each need to register. `Library::currentScope` was
called once per module, but if one does not compile to a HLTO file it
uses an implementation which returns a new value for each invocation.
This was a problem in e.g., Zeek where we do not use HLTO files, but
instead generate C++ code for all files of a collection

This patch changes generated code so it only override the scope variable
embedded in generated code if it was not already set up. We need an
extra twist for the case of separately compiled linker files as tested
by `tests/hilti/hiltic/cc/separate-units.hlt`; for that we add a new
method to `Library` which allows to detect whether `currentScope` would
return a stable or generated value.