Fix visibility of C++ globals storing linker scopes.
We emitted them as weak symbols, but that wasn't right because that
way multiple HLTOs could end up using the same scope value.

The requirements for the scope symbols are: (1) inside a single HLTO,
all object files link to the same symbol; and (2) across HLTOs, they
must not be visible. Setting their visibility to `hidden` achives this.

(I checked our other weak symbols as well, and I believe they work
correctly as is.)