Fix pass removing error push/pop.
The previous implementation in the peephole optimizer relied on an
exact sequence of statements. With removal of dead writes parts of this
could have been removed, in particular a write to a local `__error`
since it was unused.

This patch moves that pass into the `FunctionBodyVisitor` which can
clean them up before making any other edits.