Track kill set for each declaration.
We use kill sets to prevent values from propagation in the CFG. An entry
in the kill set communicates that the effect of a statement does not
propagate beyond a node, typically because a previous write was
overwritten. We previously would just track statements.

This was at odds with our tracking in the in and out set which tracked
statements by declaration, and we would incorrectly model statements
which would mutate multiple values, e.g.,

    x = 0;
    y = 0;
    x = (y = 1);   // kill `x = 0`.
    x = 2;         // Kill `x = (y = 1)`.
    assert y == 1;

In this case we would have removed the statement `x = (y = 1)` since it
did not propagate beyond `x = 2` and was never read.

With this patch we track separate kill sets for each declaration, for
above example:

    x = 0;
    y = 0;
    x = (y = 1);   // kill `x: x = 0`.
    x = 2;         // Kill `x: x = (y = 1)`.
    assert y == 1;

This allows the write `y = 1` to propagate.