Add `&cxx-any-as-ptr` attribute.
This is a very niche attribute extending the host application API:
Given an external C++ function that's made available to HILTI/Spicy
through `&cxxname`, if that function has a parameter of type `any`,
HILTI normally passes the corresponding argument by *reference* to the
argument's original type. This works well if the C++ implementation
for the target function uses a template parameter to receive the
argument, which is how we use `any` inside the HILTI/Spicy runtime
libraries at various points. It however doesn't work well if we don't
want to (or cannot) use templates, for example because the function is
part of a library without a corresponding C++ template header.

For that case, this adds a new attribute `&cxx-any-as-ptr` that can be
combined (only) with `&cxxname`: it changes the passing convention for
`any` parameters to pass a *pointer* instead of a reference. The
pointer is still typed according to the value the caller passes, but
now it can be received by implementations taking a `void *`. Along
with the type information that HILTI provides to the C++ function as a
subsequent argument (which we already did; no change), the receiving
function can now for example create a standard runtime
`type_info::Value` instance for the `any` argument and work with that
generically. If HILTI is asked to generate a prototype for the
external function, it also generates that with `void *` now if the
attribute is present.

For the record, I considered changing to pointers generally, but
by-reference has some advantages for the cases where it works, for
example by being able to pass RHS values.)

If this is all hard to comprehend, see
`tests/hilti/codegen/cxx-any-as-ptr.hlt` for an example. :)

The patch also includes a small fix for generating the C++ prototype
for such a function: we missed including the type information into the
prototype as an additional argument coming with `any` parameters.