Rework how we store the fields offsets inside a struct.
So far Spicy was tracking unit field offsets by storing an additional
vector in each compiled struct that was recording begin/end position
of the fields. That was fragile because it depended on the order of
the structs' fields matching the order of unit fields exactly,
including which ones were emitted at all. That happened to work, but
was cumbersome to maintain that way. And with all the upcoming
optimization work, depending on field specifics in the final struct,
certainly isn't not a good ideas either.

This work changes the vector to a map indexed by field ID, making
maintenance and access of the information much easier and well
defined. While somewhat slower, offset tracking is already an opt-in
feature (`spicyc -Q`) that assumes performance isn't critical.

This also brings one new feature: In addition to the fields' offsets,
we now also record beginning and end offsets of the whole struct
itself. We store that in a map entry hardcoded to `self` (because
that's a reserved ID, it cannot conflict with a field name).

Tested through the existing (and updated) spicy-dump code/tests.