Run all optimizations until convergence.
This is the first step to get the optimizer to play nicely with
coercion. We run resolution/coercion after optimization, but if we want
tho optimizer to play nice with newly-coerced results, we have to run
the optimizer after the coercion as well. This way, changes in coercion
from constant propagation can make other optimizations run.